Initial Refactor to Remove Casino Elements from Racing Game
