+++ Program.cs
@@ -25,11 +25,7 @@ namespace trillbot
             "ta!"
         };
 
-        public static Dictionary<ulong,Classes.GrandPrix> games = new Dictionary<ulong, Classes.GrandPrix>(); //For future multi-game system
-        public static Dictionary<ulong,Classes.slotMachineRunner> slots = new Dictionary<ulong, Classes.slotMachineRunner>(); //Multichannel Slot Machine
-        public static Dictionary<ulong,Classes.blackjackDealer> blackjack = new Dictionary<ulong, Classes.blackjackDealer>(); //Blackjack Games
-        public static Dictionary<ulong,Classes.roulette> roulette = new Dictionary<ulong, Classes.roulette>(); //Roulette Games
-        public static Dictionary<ulong,Classes.psiball_game> psiball = new Dictionary<ulong, Classes.psiball_game>(); //Psiball Games
+        public static Dictionary<ulong,Classes.GrandPrix> games = new Dictionary<ulong, Classes.GrandPrix>(); 
 
         public async Task RunBotAsync()
         {
@@ -65,7 +61,7 @@ namespace trillbot
         public async Task RegisterCommandAsync()
         {
             _client.MessageReceived += HandleCommandAsync;
-            await _commands.AddModulesAsync(Assembly.GetEntryAssembly());
+            await _commands.AddModulesAsync(Assembly.GetEntryAssembly(), _services);
 
         }
